backend/requirements.txt
# Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Web scraping
beautifulsoup4>=4.12.2
playwright>=1.39.0
html2text>=2020.1.16
Crawl4AI>=0.4.248

# OpenAI
openai>=1.3.0
tiktoken>=0.5.1

# Storage backends
supabase>=1.0.4
pyairtable>=2.0.0

# Utilities
aiohttp>=3.8.6
requests>=2.31.0
pypdf>=3.16.0
numpy>=1.26.0
tqdm>=4.66.1